+++ changelog.md
@@ -1,4 +1,10 @@
 # Changelog
 
 ## 0.1.0
+- allow nullable boolean columns
+- add install command
+- add layouts array to page model
+
+
+## 0.1.0
 - Initial Release